// src/enums/characteristic.ts
export enum Characteristic {
	Strength = 'Strength',
	Constitution = 'Constitution',
	Dexterity = 'Dexterity',
	Intelligence = 'Intelligence',
	Wisdom = 'Wisdom',
	Charisma = 'Charisma',

	// Old ones, gotta delete
	// keeping around to satisfy typechecker as i muck about
	Might = 'Might',
	Agility = 'Agility',
	Reason = 'Reason',
	Intuition = 'Intuition',
	Presence = 'Presence',
}